requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
alembic==1.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.0.3
python-multipart==0.0.6
resend
